SOEOPSFY24-296 | explore framer motion hero multi-overlay
