+++ app/components/Container/Container.styles.ts
@@ -5,7 +5,7 @@ export const widths = {
 };
 
 export const bgColors = {
-  black: 'bg--black text-white',
+  black: 'bg-black text-white',
   white: 'bg-white text-black',
   'fog-light': 'bg-fog-light text-black',
   'red-gradient': 'bg-soe-red-gradient text-white',
