+++ app/components/FeatureHero/FeatureHero.styles.ts
@@ -1,0 +1,4 @@
+export const root = 'h-full max-h-[1200px] relative overflow-hidden';
+export const wrapper = 'mr-0 au-ml-auto flex-col lg:flex-row';
+export const gradient = 'absolute top-0 left-0 bg-feature-gradient-rl bg-feature-gradient-bt w-full h-full z-30';
+export const contentWrapper = 'lg:rs-pr-9 ml-0 *:text-white z-50';
