+++ app/components/FeatureHero/index.ts
@@ -1,0 +1,2 @@
+export * from './FeatureHero';
+export * from './FeatureHero.styles';
