--- app/components/HeroIcon/index.ts
@@ -1,2 +1,0 @@
-export * from './HeroIcon';
-export * from './HeroIcon.styles';
