+++ app/components/ShapeAnimation/ShapeAnimation.tsx
@@ -50,8 +50,8 @@ export const ShapeAnimation = () => {
   }, [pathIndex, currentPath]);
 
   return (
-    <div>
-      <motion.svg width="500px" height="100%" viewBox="0 0 1820 1700">
+    <div className="cc rs-py-2">
+      <motion.svg width="100%" height="500px" viewBox="0 0 1820 1700">
         <motion.path d={currentPath} fill="none" stroke="black" strokeWidth="3" />
       </motion.svg>
     </div>
