+++ app/components/TimelineBanner/TimelineBanner.styles.ts
@@ -1,8 +1,8 @@
 export const root = 'overflow-hidden';
 export const wrapper = 'mr-0 au-ml-auto flex-col lg:flex-row';
-export const contentWrapper = 'lg:rs-pr-9 ml-0';
+export const contentWrapper = 'lg:rs-pr-9 ml-0 flex flex-col';
 export const heading = '2xl:whitespace-pre-line -mt-01em rs-mb-2 xl:max-w-1200';
-export const superhead = 'order-first';
+export const superhead = 'order-first mb-38';
 export const body = 'max-w-[50ch] rs-mb-3 *:*:leading-snug';
 export const imageWrapper = 'group relative w-300 h-300 perspective-600';
 export const image = 'object-cover w-full h-full transform rotate-y-30 group-hocus:-rotate-y-30 skew-y-12  ease-in duration-300 rounded-[30px]';
