+++ app/components/Typography/Text.tsx
@@ -1,6 +1,6 @@
 import { ReactNode, TimeHTMLAttributes } from 'react';
 import { cnb } from 'cnbuilder';
-import { HeroIcon, type HeroIconProps, type IconType } from '../HeroIcon';
+import { HeroIcon, type HeroIconProps, type IconType } from '../images/HeroIcon';
 import * as styles from './typography.styles';
 import * as types from './typography.types';
