+++ app/components/images/HeroIcon/index.ts
@@ -1,0 +1,2 @@
+export * from './HeroIcon';
+export * from './HeroIcon.styles';
