+++ next.config.mjs
@@ -3,6 +3,7 @@ const nextConfig = {
   basePath: '/soe-centennial-nextjs',
   output: 'export', 
   reactStrictMode: true,
+  images: { unoptimized: true },
 };
 
 export default nextConfig;
