+++ tailwind.config.ts
@@ -19,13 +19,22 @@ export default {
   ],
   theme: {
     extend: {
+      colors: {
+        'soe-black': '#181818',
+      },
       backgroundImage: {
         'soe-red-gradient': 'linear-gradient(90deg, #400000 0.19%, #7A0000 52.26%, #400000 99.94%)',
+        'feature-gradient-bt': 'linear-gradient(180deg, rgba(0, 0, 0, 0.00) 75.5%, #181818 100%)',
+        'feature-gradient-rl': 'linear-gradient(270deg, rgba(23, 23, 26, 0.00) 0.11%, #181818 99.89%)',
+      },
+      fontFamily: {
+        'dm-sans': ['var(--font-dm-sans)', 'sans-serif'],
+        stanford: ['var(--font-stanford)', 'sans-serif'],
       },
     },
   },
   plugins: [
     require('@tailwindcss/container-queries'),
-    require("@xpd/tailwind-3dtransforms")
+    require('@xpd/tailwind-3dtransforms'),
   ],
 } satisfies Config;
